reverted webtrans to revision ca3e742a0157 (apart from webtrans/server/).  also migrated packages to net.openl10n.flies again
--HG--
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/events/UserLogoutEvent.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/auth/UserLogoutEvent.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/events/UserLogoutEventHandler.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/auth/UserLogoutEventHandler.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/ContentFilter.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/ContentFilter.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/events/FilterDisabledEvent.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/FilterDisabledEvent.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/events/FilterDisabledEventHandler.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/FilterDisabledEventHandler.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/events/FilterEnabledEvent.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/FilterEnabledEvent.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/events/FilterEnabledEventHandler.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/FilterEnabledEventHandler.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/FilterPresenter.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/FilterPresenter.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/FuzzyFilter.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/FuzzyFilter.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/OperatorFilter.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/OperatorFilter.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/PhraseFilter.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/PhraseFilter.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/SourcePhraseFilter.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/SourcePhraseFilter.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/TargetPhraseFilter.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/TargetPhraseFilter.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/TransFilterMessages.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/TransFilterMessages.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/TransFilterPresenter.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/TransFilterPresenter.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/TransFilterView.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/TransFilterView.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/filter/TransFilterView.ui.xml => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/filter/TransFilterView.ui.xml
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/NavigationMessages.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/table/NavigationMessages.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/ListEditorMessages.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/table/TableEditorMessages.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/ListEditorPresenter.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/table/TableEditorPresenter.java
rename : flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/ListEditorTableDefinition.java => flies-war/src/main/java/net/openl10n/flies/webtrans/client/editor/table/TableEditorTableDefinition.java
rename : flies-war/src/main/resources/net/openl10n/flies/webtrans/client/filter/TransFilterMessages.properties => flies-war/src/main/resources/net/openl10n/flies/webtrans/client/editor/filter/TransFilterMessages.properties
rename : flies-war/src/main/resources/net/openl10n/flies/webtrans/client/editor/NavigationMessages.properties => flies-war/src/main/resources/net/openl10n/flies/webtrans/client/editor/table/NavigationMessages.properties
rename : flies-war/src/main/resources/net/openl10n/flies/webtrans/client/editor/ListEditorMessages.properties => flies-war/src/main/resources/net/openl10n/flies/webtrans/client/editor/table/TableEditorMessages.properties

+++ flies-war/src/main/java/net/openl10n/flies/webtrans/client/DocumentListPresenter.java
@@ -12,15 +12,16 @@ import net.customware.gwt.presenter.client.place.Place;
 import net.customware.gwt.presenter.client.place.PlaceRequest;
 import net.customware.gwt.presenter.client.widget.WidgetDisplay;
 import net.customware.gwt.presenter.client.widget.WidgetPresenter;
+
 import net.openl10n.flies.common.TransUnitCount;
 import net.openl10n.flies.webtrans.client.editor.HasTransUnitCount;
+import net.openl10n.flies.webtrans.client.editor.filter.ContentFilter;
 import net.openl10n.flies.webtrans.client.events.DocumentSelectionEvent;
 import net.openl10n.flies.webtrans.client.events.DocumentSelectionHandler;
 import net.openl10n.flies.webtrans.client.events.NotificationEvent;
 import net.openl10n.flies.webtrans.client.events.TransUnitUpdatedEvent;
 import net.openl10n.flies.webtrans.client.events.TransUnitUpdatedEventHandler;
 import net.openl10n.flies.webtrans.client.events.NotificationEvent.Severity;
-import net.openl10n.flies.webtrans.client.filter.ContentFilter;
 import net.openl10n.flies.webtrans.client.rpc.CachingDispatchAsync;
 import net.openl10n.flies.webtrans.shared.model.DocumentId;
 import net.openl10n.flies.webtrans.shared.model.DocumentInfo;
@@ -30,7 +31,6 @@ import net.openl10n.flies.webtrans.shared.rpc.GetDocumentList;
 import net.openl10n.flies.webtrans.shared.rpc.GetDocumentListResult;
 import net.openl10n.flies.webtrans.shared.rpc.GetProjectStatusCount;
 import net.openl10n.flies.webtrans.shared.rpc.GetProjectStatusCountResult;
-
 
 import com.allen_sauer.gwt.log.client.Log;
 import com.google.gwt.event.logical.shared.HasSelectionHandlers;
@@ -231,7 +231,7 @@ public class DocumentListPresenter extends WidgetPresenter<DocumentListPresenter
       }
       else
       {
-         long value = (translated * 100) / (fuzzy + untranslated + translated);
+         long value = (long) ((translated * 100) / (fuzzy + untranslated + translated));
          return value;
       }
